refactor: view 관련 객체 view 패키지로 이동
